backend/requirements.txt
fastapi
pymongo
motor
uvicorn
python-dotenv